+++ ScenarioGUI/gui_classes/gui_structure_classes/int_box.py
@@ -10,6 +10,7 @@ import PySide6.QtCore as QtC  # type: ignore
 import PySide6.QtWidgets as QtW  # type: ignore
 
 import ScenarioGUI.global_settings as globs
+from .functions import check_and_set_max_min_values
 
 from ...utils import set_default_font
 from .option import Option
@@ -110,6 +111,7 @@ class IntBox(Option):
         -------
         None
         """
+        check_and_set_max_min_values(self.widget, value, self.maximal_value, self.minimal_value)
         self.widget.setValue(value)
 
     def _init_links(self) -> None: